requirements.txt
assemblyai==0.45.5
asyncpg==0.29.0
beautifulsoup4==4.12.3
feedparser==6.0.11
fastapi==0.120.0
langchain==0.2.17
langchain-openai==0.1.21
langgraph==0.1.4
langsmith==0.1.116
httpx<0.28
mypy==1.18.2
openai==1.52.0
psycopg[binary,pool]==3.2.2
pytest==8.4.2
python-dotenv==1.1.1
requests==2.32.5
ruff==0.14.3
tiktoken==0.7.0
uvicorn==0.38.0
